backend/pyproject.toml
[project]
name = "road-damage-backend"
version = "0.1.0"
description = "Road Damage Detection System Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.10.0",
    "pillow>=10.0.0",
    "orjson>=3.10.0",
    "ultralytics>=8.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "neo4j>=5.0.0",
    "reportlab>=4.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "httpx>=0.27.0",
]
